requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
boto3==1.12.16
botocore==1.15.16
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9
distlib==0.3.0
dj-database-url==0.5.0
Django==2.0.6
django-ckeditor==5.9.0
django-cors-headers==2.2.0
django-filter==2.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.8.2
dnspython==1.16.0
docutils==0.15.2
filelock==3.0.12
funcy==1.14
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
mysqlclient==1.4.6
Pillow==6.2.1
pipenv==2018.11.26
protobuf==3.6.1
psycopg2==2.8.4
pycparser==2.20
pylint==2.4.4
python-dateutil==2.8.1
pytz==2018.4
requests==2.22.0
s3transfer==0.3.3
scrypt==0.8.13
six==1.14.0
sqlparse==0.3.0
toolz==0.10.0
typed-ast==1.4.1
urllib3==1.25.6
virtualenv==20.0.4
virtualenv-clone==0.5.3
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.0.0